% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggenom.R
\name{ggenom}
\alias{ggenom}
\title{Create a Base ggplot Object for Genomic Data}
\usage{
ggenom(data, mapping = NULL, plot = NULL, ...)
}
\arguments{
\item{data}{A data frame or tibble containing the genomic data to be plotted.}

\item{mapping}{Default list of aesthetic mappings to be used for the plot. Must be created using \code{\link[ggplot2:aes]{ggplot2::aes()}} or \code{\link[ggplot2:aes_]{ggplot2::aes_string()}}. Default is \code{NULL}.}

\item{plot}{A character string specifying a predefined plot type. Options include:
\itemize{
\item \code{"telplot"}: Sets mapping to \code{aes(x = Chromosome, y = begin_telo_start, yend = Length)}.
}}

\item{...}{Other arguments passed on to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}, such as \code{environment}.}
}
\value{
A ggplot object that can be further customized with additional layers, themes, and scales.
}
\description{
This function initializes a ggplot2 object for genomic data, allowing for the seamless integration of additional layers and customization.
It is designed as a lightweight wrapper around \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} with additional flexibility for genomic data visualization.
}
\details{
The \code{ggenom} function serves as a starting point for creating genomic plots. By wrapping the \code{ggplot} function, it simplifies workflows in genomic visualization
while maintaining full compatibility with ggplot2 functionality.
}
\examples{
library(ggplot2)

# Example dataset
df <- data.frame(
  Chromosome = factor(1:3),
  begin_telo_start = c(100, 200, 300),
  Length = c(1000, 1500, 1200)
)

# Using the plot argument
ggenom(df, plot = "telplot") +
  geom_telplot()

}
\seealso{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}, \code{\link[ggplot2:aes]{ggplot2::aes()}}
}
