% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggideo_diploid.R
\name{ggideo_diploid}
\alias{ggideo_diploid}
\title{Generate a diploid haplotype phased ideogram plot from two fasta file}
\usage{
ggideo_diploid(
  path_hap1_fasta,
  path_hap2_fasta,
  combined_hap_fasta = NULL,
  chr_names = "Chr",
  string_remove = "_RagTag",
  string_hap = "hap\\\\d+",
  tel_start_seq = "CCCTAAA",
  tel_end_seq = "TTTAGGG",
  size_windows = 1e+06,
  min_tel_count = 25,
  sample_name = NULL,
  title_plot = NULL,
  title_x_axis = NULL,
  title_y_axis = "Chromosome Length",
  title_legend = "Telomere Size (bp)",
  color_hap1 = "dodgerblue2",
  color_hap2 = "orangered",
  name_hap1 = "Haplotype 1",
  name_hap2 = "Haplotype 2",
  size_chr = 6,
  color_tel = "black",
  shape_tel = 16,
  scale_y = 1e-06,
  suffix_y_scale = "Mb",
  pos_legend = "bottom",
  size_legend = 0.25,
  size_text = 6
)
}
\arguments{
\item{path_hap1_fasta}{A character string specifying the file path to the fasta file containing the haplotype 1 genome sequence.}

\item{path_hap2_fasta}{A character string specifying the file path to the fasta file containing the haplotype 2 genome sequence.}

\item{combined_hap_fasta}{A character string specifying the file path to a combined FASTA file containing both haplotypes. Default is NULL.}

\item{chr_names}{A character string specifying the pattern to identify chromosome names in the data. Default is "Chr".}

\item{string_remove}{A character string to specify a substring to remove from chromosome names. Default is "_RagTag".}

\item{string_hap}{A character string specifying the pattern to identify haplotype names in the data. Default is "hap\d+".}

\item{tel_start_seq}{A character string representing the telomeric repeat sequence at the start of the chromosome. Default is "CCCTAAA".}

\item{tel_end_seq}{A character string representing the telomeric repeat sequence at the end of the chromosome. Default is "TTTAGGG".}

\item{size_windows}{A numeric value specifying the window size (in base pairs) for counting telomeric repeats. Default is \code{1e6} (1 Mb).}

\item{min_tel_count}{A numeric value representing the minimum number of telomeric repeats required to retain telomeric counts in the plot. Default is 25.}

\item{sample_name}{A character string specifying the name of the sample for labeling purposes. Default is \code{NULL}.}

\item{title_plot}{A character string specifying the title of the plot. Default is \code{NULL}.}

\item{title_x_axis}{A character string specifying the title of the x axis. Default is \code{NULL}.}

\item{title_y_axis}{A character string specifying the title of the y axis. Default is "Chromosome Length".}

\item{title_legend}{A character string specifying the title of the legend. Default is "Telomere Presence".}

\item{color_hap1}{A character string specifying the color of the chromosome segments. Default is "dodgerblue2".}

\item{color_hap2}{A character string specifying the color of the chromosome segments. Default is "orangered".}

\item{name_hap1}{A character string specifying the name of the haplotype. Default is "Haplotype 1".}

\item{name_hap2}{A character string specifying the name of the haplotype. Default is "Haplotype 2".}

\item{size_chr}{A numeric value specifying the size (linewidth) of the chromosome segments. Default is 6.}

\item{color_tel}{A character string specifying the color of the telomere points. Default is "black".}

\item{shape_tel}{A numeric value specifying the shape of the telomere points. Default is 16 (filled circle).}

\item{scale_y}{A numeric value for scaling the y-axis. Default is \code{1e-6} to convert lengths to megabases (Mb).}

\item{suffix_y_scale}{A character string to append to the y-axis labels, usually a unit like "Mb". Default is "Mb".}

\item{pos_legend}{A character string specifying the position of the legend. Default is "bottom".}

\item{size_legend}{A numeric value specifying the size of the legend keys. Default is 0.25.}

\item{size_text}{A numeric value specifying the size of text for the entire figure. Adjusts \code{base_size} ggplot function. Default is 6.}
}
\value{
A list containing:
\describe{
\item{genomic.table}{A data frame with chromosome, length, and telomere information.}
\item{ideogram}{A \code{ggplot} object representing the ideogram plot.}
}
}
\description{
This function creates an ideogram plot based on a genome sequence provided in a FASTA file. It reads the genome,
filters for specified chromosomes, calculates telomeric repeat counts, and visualizes the chromosome lengths with telomere regions marked.
The output includes a data table and a \code{ggplot} generated ideogram plot.
}
\examples{
# Example usage
# ggideo_diploid(path_hap1_fasta = "path/to/genome.fasta", path_hap2_fasta = "path/to/genome.fasta", sample_name = "Sample_1", title_plot = "Genome Ideogram")

}
