% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggread.fasta.R
\name{ggread.fasta}
\alias{ggread.fasta}
\title{Read a FASTA File into a DNAStringSet}
\usage{
ggread.fasta(path_to_fasta)
}
\arguments{
\item{path_to_fasta}{A character string specifying the path to the FASTA file.}
}
\value{
A \code{DNAStringSet} object containing the DNA sequences from the FASTA file.
}
\description{
This function reads a DNA sequence from a FASTA file and returns it as a \code{DNAStringSet} object.
}
\details{
This function serves as a wrapper around the \code{readDNAStringSet} function from the Biostrings package,
to maintain syntax when loading DNA sequences from FASTA files.
}
\examples{
\dontrun{
# Example usage:
dna_sequences <- ggread.fasta("path/to/your/fasta_file.fasta")
}

}
\seealso{
\code{\link[Biostrings:XStringSet-io]{Biostrings::readDNAStringSet()}}
}
