% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_string_chr.R
\name{remove_string_chr}
\alias{remove_string_chr}
\title{Remove Specific String from "Chromosome" Column}
\usage{
remove_string_chr(data, remove_string = "_RagTag")
}
\arguments{
\item{data}{A data frame containing a column named "Chromosome".}

\item{remove_string}{A character string to be removed from the "Chromosome" column.
Default is "_RagTag".}
}
\value{
A data frame with the specified string removed from the "Chromosome" column.
}
\description{
This function removes a specified string from the "Chromosome" column of a data frame.
It is helpful for cleaning or standardizing chromosome labels by eliminating unnecessary substrings.
}
\examples{
# Example usage
data <- data.frame(Chromosome = c("Chr1_RagTag", "Chr2_RagTag", "Chr3"))
remove_string_chr(data, remove_string = "_RagTag")

}
