% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telomere_plotting_table.R
\name{telomere_plotting_table}
\alias{telomere_plotting_table}
\title{Create a Table for Telomere Plotting}
\usage{
telomere_plotting_table(
  genome,
  chr_names = "Chr",
  string_remove = "_RagTag",
  tel_start_seq = "CCCTAAA",
  tel_end_seq = "TTTAGGG",
  size_windows = 1e+06,
  min_tel_count = 25,
  sample_name = NULL
)
}
\arguments{
\item{genome}{A fasta file read in by \code{DNAStringSet} to be analyzed.}

\item{chr_names}{A character string indicating the prefix for chromosome names (e.g., \code{"Chr"}). Default is \code{"Chr"}.}

\item{string_remove}{A character string to remove from chromosome names. Default is \code{"_RagTag"}.}

\item{tel_start_seq}{A character string representing the telomere start sequence. Default is \code{"CCCTAAA"}.}

\item{tel_end_seq}{A character string representing the telomere end sequence. Default is \code{"TTTAGGG"}.}

\item{size_windows}{A numeric value specifying the window size (in base pairs) for telomere repeat counting. Default is \code{1e6}.}

\item{min_tel_count}{A numeric value specifying the minimum telomere repeat count to include in the final table. Default is \code{25}.}

\item{sample_name}{A character string representing the sample name. Default is \code{NULL}.}
}
\value{
A data frame with columns for chromosome names, chromosome lengths, and telomere repeat counts. Chromosome names are properly formatted and ordered.
}
\description{
This function generates a table of chromosomes, their lengths, and the associated telomere data for plotting purposes.
The table includes genomic information, telomere repeat counts, and proper formatting to ensure correct chromosome ordering and visualization.
}
\details{
This function performs several tasks:
\itemize{
\item Filters chromosome data based on the provided chromosome name prefix.
\item Counts telomeric repeats in windows of the genome using specified start and end telomere sequences.
\item Filters out chromosomes with fewer telomere repeats than the specified threshold (\code{min_tel_count}).
\item Formats chromosome names (removes specific strings, removes leading zeros, and ensures correct ordering).
}

The resulting table can be used for plotting telomere data and chromosome lengths.
}
\examples{
\dontrun{
library(Biostrings)
genome <- readDNAStringSet("path/to/genome.fasta")
telomere_plotting_table(genome, chr_names = "Chr", string_remove = "_RagTag")
}

}
\seealso{
\code{\link[=select_chr]{select_chr()}}, \code{\link[=telomere_repeat_number]{telomere_repeat_number()}}, \code{\link[=genome_table]{genome_table()}}, \code{\link[=remove_string_chr]{remove_string_chr()}}, \code{\link[=remove_lead_0s]{remove_lead_0s()}}, \code{\link[=remove_trailing]{remove_trailing()}}
}
